// src/app/interfaces/cast-response.ts
export interface CastResponse {
  id: number;
  cast: Cast[];
  crew: Cast[];
}

export interface Cast {
  adult: boolean;
  gender: number;
  id: number;
  known_for_department: Department;
  name: string;
  original_name: string;
  popularity: number;
  profile_path: null | string;
  cast_id?: number;
  character?: string;
  credit_id: string;
  order?: number;
  department?: Department;
  job?: string;
}

export enum Department {
  Acting = 'Acting',
  Camera = 'Camera',
  Crew = 'Crew',
  Directing = 'Directing',
  Production = 'Production',
  Writing = 'Writing',
}